Bug fix.
See accompanying test case. In the case where tail had a length of 0, we didn't
push it into ops. But formatBoolAttrs could change the length of tail, and then
it indeed needs to be in ops. The simplest solution is to just push tail
initially, and then let delta.compact handle this in the end, which we are
already calling anyways.