Remove normalize as it's no longer needed.
We used to use normalize to handle our transition from representing inserts as
strings and retains as pairs. Those data types are no longer in the system, so
we don't need normalize. However, normalize was actually deep copying ops when
called form compact, which was called from the Delta constructor. So we now deep
copy ops in the Delta constructor directly, and can completely remove normalize.