Replace underscore with lodash.
We can use a custom build (or even just the preset underscore build) if we want
to optimize, or if any bugs arise from this change. But, the test suite passes
after the change, and lodash is small enough that it seems like premature
optimization at this point. See
http://kitcambridge.be/blog/custom-builds-in-lo-dash-2-dot-0/ for details on
custom builds and when to use them.